requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI & ML
openai==1.3.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-multipart==0.0.6

# Audio Processing
sarvamai==0.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1